Move state files
